libvirt: proper monitoring of live migration progress
The current live migration code simply invokes migrateToURI
and waits for it to finish, or raise an exception. It considers
all exceptions to mean the live migration aborted and the VM is
still running on the source host. This is totally bogus, as there
are a number of reasons why an error could be raised from the
migrateToURI call. There are at least 5 different scenarios for
what the VM might be doing on source + dest host upon error.
The migration might even still be going on, even if after the
error has occurred.

A more reliable way to deal with this is to actively query
libvirt for the domain job status. This gives an indication
of whether the job is completed, failed or cancelled. Even
with that though, there is a need for a few heuristics to
distinguish some of the possible error scenarios.

This change to do active monitoring of the live migration process
also opens the door for being able to tune live migration on the
fly to adjust max downtime or bandwidth to improve chances of
getting convergence, or to automatically abort it after too much
time has elapsed instead of letting it carry on until the end of
the universe. This change merely records memory transfer progress
and leaves tuning improvements to a later date.

Closes-bug: #1414065
Change-Id: I6fcbfa31a79c7808c861bb3a84b56bd096882004
(cherry picked from commit 7dd6a4a19311136c02d89cd2afd97236b0f4cc27)

Conflicts:
	nova/tests/virt/libvirt/fakelibvirt.py
	nova/tests/virt/libvirt/test_driver.py
	nova/tests/virt/libvirt/test_host.py
	nova/virt/libvirt/driver.py
	nova/virt/libvirt/host.py

NOTE(apporc), conflicts are mainly because of:
1. commit 89cd6a0c493e26b5a9e017c99d731464292abbaf, all tests have
been moved to nova/tests/unit.
2. commit 4b9bec3a1c819e7006af6bfa1de6928bde91b2c7, mock libvirt
class refactor.
3. commit 069ad9a5b4b2cc9b49b21b09b0667909886961cf, new 'Host'
class as a cap of libvirt connection.